requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
mysql==0.0.3
mysql-connector-python==9.4.0
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
